+++ sdk/eventhubs/azure-messaging-eventhubs/test/ut/eventhubs_test_base.hpp
@@ -12,7 +12,7 @@
 enum class AuthType
 {
   Key,
-  //  ConnectionString,
+  ConnectionString,
   Emulator,
 };
 
@@ -48,14 +48,19 @@ protected:
         return std::make_unique<Azure::Messaging::EventHubs::ConsumerClient>(
             eventHubNamespace, eventHubName, GetTestCredential(), eventConsumerGroup, options);
       }
+      case AuthType::ConnectionString: {
+        std::string connectionString = GetEnv("EVENTHUB_CONNECTION_STRING");
+        return std::make_unique<Azure::Messaging::EventHubs::ConsumerClient>(
+            connectionString, eventHubName, eventConsumerGroup, options);
+      }
       case AuthType::Emulator: {
-        //   return std::make_unique<Azure::Messaging::EventHubs::ConsumerClient>(
-        //       "Endpoint=sb://localhost:5672/"
-        //       ";SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=abcdefabcdef;"
-        //       "UseDevelopmentEmulator=true",
-        //       "eh1",
-        //       "$default",
-        //       options);
+        return std::make_unique<Azure::Messaging::EventHubs::ConsumerClient>(
+            "Endpoint=sb://localhost:5672/"
+            ";SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=abcdefabcdef;"
+            "UseDevelopmentEmulator=true",
+            "eh1",
+            "$default",
+            options);
       }
     }
     return nullptr;
@@ -79,13 +84,20 @@ protected:
             eventHubNamespace, eventHubName, GetTestCredential(), options);
         break;
       }
+      case AuthType::ConnectionString: {
+        std::string connectionString = GetEnv("EVENTHUB_CONNECTION_STRING");
+        producer = std::make_unique<Azure::Messaging::EventHubs::ProducerClient>(
+            connectionString, eventHubName, options);
+        break;
+      }
       case AuthType::Emulator: {
-        //   producer = std::make_unique<Azure::Messaging::EventHubs::ProducerClient>(
-        //       "Endpoint=sb://localhost:5672/"
-        //       ";SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=abcdefabcdef;"
-        //       "UseDevelopmentEmulator=true",
-        //       "eh1",
-        //       options);
+        producer = std::make_unique<Azure::Messaging::EventHubs::ProducerClient>(
+            "Endpoint=sb://localhost:5672/"
+            ";SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=abcdefabcdef;"
+            "UseDevelopmentEmulator=true",
+            "eh1",
+            options);
+        break;
       }
     }
     return producer;